refactor: 불필요한 코드 삭제 및 수정
