+++ backend/pokerogue/src/main/java/com/pokerogue/helper/item/repository/ItemRepository.java
@@ -1,5 +1,6 @@
-package com.pokerogue.helper.item;
+package com.pokerogue.helper.item.repository;
 
+import com.pokerogue.helper.item.data.Item;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface ItemRepository extends MongoRepository<Item,String> {
